refactor: inline bun secrets in storage
